+++ src/compas_timber/connections/joint.py
@@ -1,9 +1,7 @@
 from itertools import combinations
 
 from compas.geometry import Point
-from compas.geometry import angle_vectors
 from compas.geometry import distance_point_line
-from compas.geometry import intersection_line_line
 from compas_model.interactions import Interaction
 
 from .solver import JointTopology
@@ -29,15 +27,28 @@ class Joint(Interaction):
         A list of features that were added to the beams by this joint.
     attributes : dict
         A dictionary of additional attributes for this joint.
-
+    topology : literal, one of JointTopology.TOPO_UNKNOWN, JointTopology.TOPO_L, JointTopology.TOPO_T, JointTopology.TOPO_X, JointTopology.TOPO_I
+        The topology by which the two elements connected with this joint interact.
+    location : :class:`~compas.geometry.Point`
+        The estimated location of the interaction point of the two elements connected with this joint.
     """
 
     SUPPORTED_TOPOLOGY = JointTopology.TOPO_UNKNOWN
     MIN_ELEMENT_COUNT = 2
     MAX_ELEMENT_COUNT = 2
 
-    def __init__(self, **kwargs):
+    def __init__(self, topology=None, location=None, **kwargs):
         super(Joint, self).__init__(name=self.__class__.__name__)
+        self._topology = topology if topology is not None else JointTopology.TOPO_UNKNOWN
+        self._location = location or Point(0, 0, 0)
+
+    @property
+    def topology(self):
+        return self._topology
+
+    @property
+    def location(self):
+        return self._location
 
     @property
     def elements(self):
@@ -161,106 +172,3 @@ class Joint(Interaction):
         for pair in combinations(self.elements, 2):
             interactions.append((pair[0], pair[1]))
         return interactions
-
-    @staticmethod
-    def get_face_most_towards_beam(beam_a, beam_b, ignore_ends=True):
-        """Of all the faces of `beam_b`, returns the one whose normal most faces `beam_a`.
-
-        This is done by calculating the inner-product of `beam_a`'s centerline which each of the face normals of `beam_b`.
-        The face with the result closest to 1 is chosen.
-
-        Parameters
-        ----------
-        beam_a : :class:`~compas_timber.parts.Beam`
-            The beam that attaches with one of its ends to `beam_b`.
-        beam_b : :class:`~compas_timber.parts.Beam`
-            The other beam.
-        ignore_ends : bool, optional
-            If True, the faces at each end of `beam_b` are ignored.
-
-        Returns
-        -------
-        tuple(face_index, :class:`~compas.geometry.Frame`)
-            Tuple containing the index of the chosen face and a frame at the center of if.
-
-        """
-        face_dict = Joint._beam_side_incidence(beam_a, beam_b, ignore_ends)
-        face_index = max(face_dict, key=face_dict.get)  # type: ignore
-        return face_index, beam_b.faces[face_index]
-
-    @staticmethod
-    def get_face_most_ortho_to_beam(beam_a, beam_b, ignore_ends=True):
-        """Of all the faces of `beam_b`, returns the one whose normal is most orthogonal to `beam_a`.
-
-        This is done by calculating the inner-product of `beam_a`'s centerline which each of the face normals of `beam_b`.
-        The face with the result closest to 0 is chosen.
-
-        Parameters
-        ----------
-        beam_a : :class:`~compas_timber.parts.Beam`
-            The beam that attaches with one of its ends to `beam_b`.
-        beam_b : :class:`~compas_timber.parts.Beam`
-            The other beam.
-        ignore_ends : bool, optional
-            If True, the faces at each end of `beam_b` are ignored.
-
-        Returns
-        -------
-        tuple(face_index, :class:`~compas.geometry.Frame`)
-            Tuple containing the index of the chosen face and a frame at the center of if.
-
-        """
-        face_dict = Joint._beam_side_incidence(beam_a, beam_b, ignore_ends)
-        face_index = min(face_dict, key=face_dict.get)  # type: ignore
-        return face_index, beam_b.faces[face_index]
-
-    @staticmethod
-    def _beam_side_incidence(beam_a, beam_b, ignore_ends=True):
-        """Returns a map of face indices of beam_b and the angle of their normal with beam_a's centerline.
-
-        This is used to find a cutting plane when joining the two beams.
-
-        Parameters
-        ----------
-        beam_a : :class:`~compas_timber.parts.Beam`
-            The beam that attaches with one of its ends to the side of beam_b.
-        beam_b : :class:`~compas_timber.parts.Beam`
-            The other beam.
-        ignore_ends : bool, optional
-            If True, only the first four faces of `beam_b` are considered. Otherwise all faces are considered.
-
-        Examples
-        --------
-        >>> face_angles = Joint.beam_side_incidence(beam_a, beam_b)
-        >>> closest_face_index = min(face_angles, key=face_angles.get)
-        >>> cutting_plane = beam_b.faces[closest_face_index]
-
-        Returns
-        -------
-        dict(int, float)
-            A map of face indices of beam_b and their respective angle with beam_a's centerline.
-
-        """
-        # find the orientation of beam_a's centerline so that it's pointing outward of the joint
-        # find the closest end
-        p1x, _ = intersection_line_line(beam_a.centerline, beam_b.centerline)
-        if p1x is None:
-            raise AssertionError("No intersection found")
-
-        end, _ = beam_a.endpoint_closest_to_point(Point(*p1x))
-
-        if end == "start":
-            centerline_vec = beam_a.centerline.vector
-        else:
-            centerline_vec = beam_a.centerline.vector * -1
-
-        if ignore_ends:
-            beam_b_faces = beam_b.faces[:4]
-        else:
-            beam_b_faces = beam_b.faces
-
-        face_angles = {}
-        for face_index, face in enumerate(beam_b_faces):
-            face_angles[face_index] = angle_vectors(face.normal, centerline_vec)
-
-        return face_angles